refactored WebPlatformPlayer to decouple from webview and allow custom implementations; Added a WebPTXPlayer implementation
