+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/LiveAppSDKLogic.ts
