+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/WalletAPIWebview.tsx
@@ -1,6 +1,6 @@
 import { WebviewTag } from "electron";
 import * as remote from "@electron/remote";
-import React, { useCallback, useEffect, useMemo, useRef } from "react";
+import React, { ForwardedRef, useCallback, useEffect, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 
@@ -19,7 +19,6 @@ import { getEnv } from "@ledgerhq/live-common/env";
 
 import { openModal } from "../../actions/modals";
 import { updateAccountWithUpdater } from "../../actions/accounts";
-import TrackPage from "../../analytics/TrackPage";
 import useTheme from "../../hooks/useTheme";
 import { flattenAccountsSelector } from "../../reducers/accounts";
 import BigSpinner from "../BigSpinner";
@@ -28,16 +27,11 @@ import { OperationDetails } from "~/renderer/drawers/OperationDetails";
 import SelectAccountAndCurrencyDrawer from "~/renderer/drawers/DataSelector/SelectAccountAndCurrencyDrawer";
 import { track } from "~/renderer/analytics/segment";
 import { shareAnalyticsSelector } from "~/renderer/reducers/settings";
-import TopBar from "./TopBar";
-import { TopBarConfig } from "./type";
-import { Container, Wrapper, Loader } from "./styled";
+import { Loader } from "./styled";
+import { useCombinedRefs } from "./helpers";
 
 const wallet = { name: "ledger-live-desktop", version: __APP_VERSION__ };
 const tracking = trackingWrapper(track);
-
-export type WebPlatformPlayerConfig = {
-  topBarConfig?: TopBarConfig;
-};
 
 function useUiHook(manifest: AppManifest): Partial<UiHook> {
   const { pushToast } = useToasts();
@@ -160,10 +154,14 @@ function useWalletAPIUrl({ manifest, inputs }: Omit<Props, "onClose">) {
   );
 }
 
-function useWebView({ manifest, inputs }: Pick<Props, "manifest" | "inputs">) {
+function useWebView(
+  { manifest, inputs }: Pick<Props, "manifest" | "inputs">,
+  ref: ForwardedRef<WebviewTag>,
+) {
   const accounts = useSelector(flattenAccountsSelector);
 
-  const webviewRef = useRef<WebviewTag>(null);
+  const innerRef = React.useRef();
+  const webviewRef = useCombinedRefs<WebviewTag>(ref, innerRef);
   const uiHook = useUiHook(manifest);
   const url = useWalletAPIUrl({ manifest, inputs });
   const shareAnalytics = useSelector(shareAnalyticsSelector);
@@ -185,6 +183,7 @@ function useWebView({ manifest, inputs }: Pick<Props, "manifest" | "inputs">) {
         }
       },
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const { widgetLoaded, onLoad, onReload, onMessage } = useWalletAPIServer({
@@ -216,6 +215,7 @@ function useWebView({ manifest, inputs }: Pick<Props, "manifest" | "inputs">) {
     // cf. https://gist.github.com/codebytere/409738fcb7b774387b5287db2ead2ccb
     // @ts-expect-error: missing typings for api
     window.api.openWindow(id);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {
@@ -237,6 +237,7 @@ function useWebView({ manifest, inputs }: Pick<Props, "manifest" | "inputs">) {
         webview.removeEventListener("dom-ready", handleDomReady);
       }
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [handleDomReady, handleMessage, onLoad]);
 
   const webviewStyle = useMemo(() => {
@@ -254,29 +255,21 @@ function useWebView({ manifest, inputs }: Pick<Props, "manifest" | "inputs">) {
 
 interface Props {
   manifest: AppManifest;
-  onClose?: () => void;
   inputs?: Record<string, string>;
-  config?: WebPlatformPlayerConfig;
 }
 
-export function WebView({ manifest, onClose, inputs = {}, config }: Props) {
-  const { webviewRef, webviewStyle, url, widgetLoaded, onReload } = useWebView({
-    manifest,
-    inputs,
-  });
+export const WalletAPIWebview = React.forwardRef<WebviewTag, Props>(
+  ({ manifest, inputs = {} }, ref) => {
+    const { webviewRef, webviewStyle, url, widgetLoaded } = useWebView(
+      {
+        manifest,
+        inputs,
+      },
+      ref,
+    );
 
-  return (
-    <Container>
-      <TrackPage category="Platform" name="App" appId={manifest.id} params={inputs} />
-      <TopBar
-        manifest={manifest}
-        onReload={onReload}
-        onClose={onClose}
-        webviewRef={webviewRef}
-        config={config?.topBarConfig}
-      />
-
-      <Wrapper>
+    return (
+      <>
         <webview
           src={url.toString()}
           ref={webviewRef}
@@ -302,7 +295,9 @@ export function WebView({ manifest, onClose, inputs = {}, config }: Props) {
             <BigSpinner data-test-id="live-app-loading-spinner" size={50} />
           </Loader>
         ) : null}
-      </Wrapper>
-    </Container>
-  );
-}
+      </>
+    );
+  },
+);
+
+WalletAPIWebview.displayName = "WalletAPIWebview";
