+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/styled.tsx
@@ -1,21 +1,7 @@
 // Need react for the Box import, if not typescript is not happy
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import React from "react";
 import styled from "styled-components";
-import Box from "../Box";
 
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  height: 100%;
-`;
-
-export const Wrapper = styled(Box).attrs(() => ({
-  flex: 1,
-}))`
-  position: relative;
-`;
 export const Loader = styled.div`
   display: flex;
   align-items: center;
