+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/type.js
@@ -1,0 +1,14 @@
+// @flow
+
+export type Manifest = {
+  name: string,
+  url: URL,
+  icon?: string,
+};
+
+export type TopBarConfig = {
+  shouldDisplayName?: boolean,
+  shouldDisplayInfo?: boolean,
+  shouldDisplayClose?: boolean,
+  shouldDisplayNavigation?: boolean,
+};
