+++ apps/ledger-live-desktop/src/renderer/components/WebPTXPlayer/index.tsx
@@ -1,10 +1,10 @@
 import { WebviewTag } from "electron";
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
-import React, { useCallback, useRef, useState } from "react";
+import React, { useCallback, useState } from "react";
 import styled from "styled-components";
 
 import Web3AppWebview from "../Web3AppWebview";
-import { TopBar, TopBarConfig } from "./TopBar";
+import { TopBar } from "./TopBar";
 import Box from "../Box";
 
 export const Container = styled.div`
@@ -20,18 +20,12 @@ export const Wrapper = styled(Box).attrs(() => ({
   position: relative;
 `;
 
-export type WebPlatformPlayerConfig = {
-  topBarConfig?: TopBarConfig;
-};
-
 interface Props {
   manifest: LiveAppManifest;
   inputs?: Record<string, string>;
-  onClose?: () => void;
-  config?: WebPlatformPlayerConfig;
 }
 
-export default function WebPlatformPlayer({ manifest, inputs, onClose, config }: Props) {
+export default function WebPTXPlayer({ manifest, inputs }: Props) {
   const [webview, setWebview] = useState<WebviewTag>(null);
 
   const webviewRef = useCallback(node => {
@@ -41,12 +35,7 @@ export default function WebPlatformPlayer({ manifest, inputs, onClose, config }:
   return (
     <Container>
       <Wrapper>
-        <TopBar
-          manifest={manifest}
-          onClose={onClose}
-          webview={webview}
-          config={config?.topBarConfig}
-        />
+        {webview ? <TopBar manifest={manifest} webview={webview} /> : null}
         <Web3AppWebview manifest={manifest} inputs={inputs} ref={webviewRef} />
       </Wrapper>
     </Container>
